rename tof_1d
